pyproject.toml
[tool.poetry]
name = "dealsnoop"
version = "0.1.0"
description = ""
authors = [
    "iRedSC"
]
readme = "README.md"

packages = [{ include = "dealsnoop", from = "src" }]



[tool.poetry.dependencies]
python = ">=3.12"
beautifulsoup4 = ">=4.13.5,<5.0.0"
selenium = ">=4.35.0,<5.0.0"
dotenv = ">=0.9.9,<0.10.0"
openai = ">=1.107.0,<2.0.0"
webdriver-manager = ">=4.0.2,<5.0.0"
discord-py = ">=2.6.3,<3.0.0"
chromedriver-autoinstaller = ">=0.6.4,<0.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
